pyproject.toml
[build-system]
requires = ["setuptools>=67.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "carpy"
description = """Consolidated aircraft recipes in Python."""
readme = "README.md"
requires-python = ">=3.9.5"
license = { file = "LICENSE.md" }
authors = [
    { name = "Yaseen Reza", email = "yaseen605@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "fsspec",
    "matplotlib",
    "numpy",
    "openpyxl",
    "pandas",
    "periodictable",
    "pytest>=7",
    "python-metar",
    "pyyaml",
    "requests",
    "sectionproperties",
    "scipy",
    "setuptools>=67.0.0",
    "sympy",
]
dynamic = ["version"]

[project.optional-dependencies]
ipynb = ["jupyter", "notebook"]

[project.urls]
homepage = "https://github.com/yaseen157/carpy"
documentation = "https://pypi.org/project/carpy/"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools.dynamic]
version = { attr = "carpy.__version__" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.csv", "*.xlsx", "*.yaml"]
